requirements.txt

streamlit
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters

transformers
sentence-transformers
sentencepiece

pandas
numpy
scikit-learn
scipy

fastapi
uvicorn

beautifulsoup4
chroma-hnswlib
chromadb

requests
aiohttp

python-dotenv
pydantic

psycopg2-binary
tqdm